メッセージをモジュール化
