+++ src/hanabi/src/FamilyMessages.ts
@@ -1,0 +1,11 @@
+const familyMessages = [
+    'こんばんはわ～',
+    '今日もかわいいね',
+    '浴衣似合ってるよ',
+    '花火綺麗だね',
+    '夏祭り楽しかったね',
+    '一緒にいられて幸せ',
+    '来年もまた来ようね',
+    '楽しいね'
+] as const;
+export default familyMessages;
